RSC: Add RscCache WebSocket support to dev server
